fix: move CloudNotificationContent to platform/cloud/notification/
Aligns with layered architecture where cloud-specific components live
under src/platform/cloud/ (peer to subscription/ and onboarding/).

Addresses review feedback:
https://github.com/Comfy-Org/ComfyUI_frontend/pull/10116#discussion_r2945375766
