+++ src/platform/cloud/notification/components/CloudNotificationContent.vue
